Move the RemovedDependencyNotation class to its own file
