+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/removals_replacement/RemovedDependencyNotation.kt
@@ -1,0 +1,10 @@
+package de.fayard.refreshVersions.core.internal.removals_replacement
+
+import de.fayard.refreshVersions.core.ModuleId
+
+internal data class RemovedDependencyNotation(
+    val dependencyNotation: String,
+    val moduleId: ModuleId.Maven,
+    val leadingCommentLines: List<String>,
+    val replacementMavenCoordinates: ModuleId.Maven?
+)
