+++ plugins/core/src/main/kotlin/de/fayard/refreshVersions/core/internal/removals_replacement/RemovedDependencyNotationsHistoryParsing.kt
@@ -3,13 +3,6 @@ package de.fayard.refreshVersions.core.internal.removals_replacement
 import de.fayard.refreshVersions.core.ModuleId
 import java.io.InputStream
 
-internal data class RemovedDependencyNotation(
-    val dependencyNotation: String,
-    val moduleId: ModuleId.Maven,
-    val leadingCommentLines: List<String>,
-    val replacementMavenCoordinates: ModuleId.Maven?
-)
-
 internal fun InputStream.parseRemovedDependencyNotationsHistory(currentRevision: Int): List<RemovedDependencyNotation> {
     return bufferedReader().lineSequence().parseRemovedDependencyNotationsHistory(currentRevision)
 }
